zad1/requirements.txt
amqp==5.1.1
asgiref==3.5.1
billiard==3.6.4.0
celery==5.2.6
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.0.4
django-environ==0.8.1
django-filter==21.1
django-seed==0.3.1
djangorestframework==3.13.1
Faker==13.11.1
greenlet==1.1.2
kombu==5.2.4
prompt-toolkit==3.0.29
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.36
sqlparse==0.4.2
toposort==1.7
vine==5.0.0
wcwidth==0.2.5
